Use size_t in mr2d_malloc
In C, the type of the number of bytes passed to `malloc` *is* of type `size_t`.

The current code ties that number of bytes to the general "Index type" (`Int`) which is defined at compile time and does fail on 32-bits systems setting `Int` as 64-bits.

Additionally, this artificially limits 64-bit-element aux matrices to 2GB instead of 16GB with 32-bits signed indices.

+++ REDIST/SRC/pgemraux.c
@@ -105,7 +105,7 @@ extern void Cpigemr2d();
 #include <stdlib.h>
 #include <assert.h>
 void *
-mr2d_malloc(Int n)
+mr2d_malloc(size_t n)
 {
   void *ptr;
   assert(n > 0);
